Test repository init state
